requirements.txt
discord.py==2.3.2
pydantic==2.4.2
requests==2.31.0

# Temporary for fixing twitter embeds.
beautifulsoup4==4.12.2
httpx==0.25.1
